cpu/stm32: group periph definitions by periph type
